Remove in-source build of `pcaudiolib` in Windows
While the in-source build makes MSVC build easier in Windows,
and that can still be done with the bundled MSVC solutions,
it does not make sense for the MinGW64 (gcc) build.
In there, building `pcaudiolib` requires extra dependencies to FAudio,
which in turn depends on SDL2.
That means there is a need to deal with dependencies already,
and the in-source build loses its advantage.

This lets CMake handle the dependency to `pcaudiolib`.
Note that the MinGW64 still does not compile properly,
and data generation still fails
due to missing binary library search paths.