Move eSpeak NG data compilation to install time
The data compilation was previously set up to be done in build time.
However, since the compilation uses the eSpeak NG binary
which when executed has to find any dynamic libraries that it depends on,
it is possible that it ends up unable to find the required libraries.
For example, in Windows, binary executables can search for DLLs in,
roughly speaking
(source: <https://docs.microsoft.com/windows/win32/dlls/dynamic-link-library-search-order):

  * The directory the binary is in,
  * System directories,
  * The current directory and
  * Directories listed in `%PATH%`.

While the executable binary might be able
to find dependencies after installation,
the install location is usually not in any of these
for the built library that lives in the build directory.
The end result is then data compilation failing.

Data compilation is moved to the install time to avoid this.