Allow custom path to eSpeak NG source
This repository contains the CMake build script for compiling eSpeak NG.
In particular, it does not contain eSpeak NG source files.
So the build script needs to find the source files somehow.
It was previously hard-coded to expect them in a sibling directory
```
|-- espeak-ng-cmake  (This repository)
|   |-- CMakeLists.txt
|-- espeak-ng  (eSpeak NG repository)
    |-- src
        |-- espeak-ng.c
```
The build script is changed to allow a different path to be specified,
with some "sensible" defaults.