Set default build options based on OS
The CMake build scripts inherits some options from Autoconf scripts
to be implemented in the future
(and to reduce the need for me to check Autoconf scripts later on).
Those options were toggled on or off by default
based on whether the build was targetting MSVC.
In particular, a minimally usable set of feature options was used
for the MSVC target.
But since the MSYS2 MinGW64 target is also experimental,
it makes sense for the minimal feature set to be used for it as well.

With this, the default build options produces a minimally usable binary.