Add new known issues into `README.md`
This CMake build of eSpeak NG is experimental.
So issues are expected.
While issues are normally better suited for the issue tracker,
since I do not have a good understanding of the codebase,
I am unable to make meaningful issues,
as illustrated by the issue descriptions.
So the issues are currently written in the `README.md`
as a mini to-do list.