Fix CMake build not searching for dependencies
CMake provides a way to create and export a package configuration
that other CMake users can use to link to `espeak-ng`.
If a `espeak-ng` build depends on `pcaudiolib`,
then the exported configuration from `espeak-ng`
should search for the package configuration exported by `pcaudiolib`,
to ensure the depenency is automatically satisfied.
Similarly for `Threads` package which is not optional.

These searches have been added to the exported package.