Copy Autoconf steps for MBROLA option
Building eSpeak NG with Autoconf provides an option to use MBROLA.
This CMake build currently lacks MBROLA support.

The steps in Autoconf has now been copied over.
However, noting the usage instructions for Windows,
the steps seem to not be applicable on Windows.
So the steps are now added, but untested.
And also unused, since this CMake build is only targetting MSYS2 MinGW64.